+++ R/parse-response.R
@@ -22,6 +22,16 @@ parse_response.default <- function(response) {
   parsed
 }
 
+#' @export
+#' @describeIn parse_response Many of the endpoints return a list of named
+#'   values for each page, which can easily be row-bound. They often contain
+#'   nested data, however, which is automatically unnested by dplyr::bind_rows.
+#'   Hence this more basic approach.
+parse_response.row_list <- function(response) {
+  purrr::list_transpose(response) %>%
+    tibble::as_tibble()
+}
+
 flatten_bind <- function(response) {
   parsed <- purrr::map(response, purrr::list_flatten)
   parsed <- purrr::list_transpose(parsed)